// src/pages/register/register.types.ts
export interface RegisterElements {
  nameInput: HTMLInputElement;
  emailInput: HTMLInputElement;
  passwordInput: HTMLInputElement;
  confirmPasswordInput: HTMLInputElement;
  registerBtn: HTMLButtonElement;
  messageBox: HTMLDivElement;
  form: HTMLFormElement;
  formSection: HTMLElement;
  nameError: HTMLElement;
  emailError: HTMLElement;
  passwordError: HTMLElement;
  confirmPasswordError: HTMLElement;
}